Allow inheritance only from imports
